refactor(A11Y): put helper functions into accessibility/index
