+++ packages/core/acessibility/accessibility-service-common.ts
@@ -1,63 +1,7 @@
 import { View } from '../ui/core/view';
 import { Page } from '../ui/page';
 import { Observable } from '../data/observable';
-import { AccessibilityBlurEventData, AccessibilityFocusChangedEventData, AccessibilityFocusEventData } from './types';
-
-const lastFocusedViewOnPageKeyName = '__lastFocusedViewOnPage';
-
-export function getLastFocusedViewOnPage(page: Page): View | null {
-	try {
-		const lastFocusedViewRef = page[lastFocusedViewOnPageKeyName] as WeakRef<View>;
-		if (!lastFocusedViewRef) {
-			return null;
-		}
-
-		const lastFocusedView = lastFocusedViewRef.get();
-		if (!lastFocusedView) {
-			return null;
-		}
-
-		if (!lastFocusedView.parent || lastFocusedView.page !== page) {
-			return null;
-		}
-
-		return lastFocusedView;
-	} catch {
-		// ignore
-	} finally {
-		delete page[lastFocusedViewOnPageKeyName];
-	}
-
-	return null;
-}
-
-export function notifyAccessibilityFocusState(view: View, receivedFocus: boolean, lostFocus: boolean): void {
-	if (!receivedFocus && !lostFocus) {
-		return;
-	}
-
-	view.notify({
-		eventName: View.accessibilityFocusChangedEvent,
-		object: view,
-		value: !!receivedFocus,
-	} as AccessibilityFocusChangedEventData);
-
-	if (receivedFocus) {
-		if (view.page) {
-			view.page[lastFocusedViewOnPageKeyName] = new WeakRef(view);
-		}
-
-		view.notify({
-			eventName: View.accessibilityFocusEvent,
-			object: view,
-		} as AccessibilityFocusEventData);
-	} else if (lostFocus) {
-		view.notify({
-			eventName: View.accessibilityBlurEvent,
-			object: view,
-		} as AccessibilityBlurEventData);
-	}
-}
+import { AccessibilityBlurEventData, AccessibilityFocusChangedEventData, AccessibilityFocusEventData } from './accessibility-types';
 
 export interface SharedA11YObservable extends Observable {
 	readonly accessibilityServiceEnabled?: boolean;
