+++ packages/core/acessibility/acessibility-properties.ts
@@ -2,7 +2,7 @@ import { booleanConverter } from '../ui/core/view-base';
 import { CssProperty, InheritedCssProperty, Property } from '../ui/core/properties';
 import { Style } from '../ui/styling/style';
 import { View } from '../ui/core/view';
-import { AccessibilityLiveRegion, AccessibilityRole, AccessibilityState, AccessibilityTrait } from './types';
+import { AccessibilityLiveRegion, AccessibilityRole, AccessibilityState, AccessibilityTrait } from './accessibility-types';
 
 function makePropertyEnumConverter<T>(enumValues: any) {
 	return (value: string): T | null => {
