+++ packages/core/acessibility/index.ios.ts
@@ -1,8 +1,12 @@
+import { ProxyViewContainer } from 'ui/proxy-view-container';
 import * as Application from '../application';
 import { View } from '../ui/core/view';
-import { ProxyViewContainer } from '../ui/proxy-view-container';
-import { AccessibilityLiveRegion, AccessibilityRole, AccessibilityState, AccessibilityTrait } from './types';
-import { notifyAccessibilityFocusState } from './accessibility-service';
+import { notifyAccessibilityFocusState } from './accessibility-common';
+import { AccessibilityLiveRegion, AccessibilityRole, AccessibilityState, AccessibilityTrait } from './accessibility-types';
+
+export * from './accessibility-css-helper';
+export * from './accessibility-types';
+export * from './fontscale-observable';
 
 function enforceArray(val: string | string[]): string[] {
 	if (Array.isArray(val)) {
@@ -35,7 +39,7 @@ function inputArrayToBitMask(values: string | string[], map: Map<string, number>
 let AccessibilityTraitsMap: Map<string, number>;
 let RoleTypeMap: Map<AccessibilityRole, number>;
 
-let nativeFocusedNotificationObserver: any;
+let nativeFocusedNotificationObserver: NSNotification;
 const uiViewToTnsView = new WeakMap<UIView, WeakRef<View>>();
 let lastFocusedView: WeakRef<View>;
 function ensureNativeClasses() {
@@ -110,108 +114,97 @@ function ensureNativeClasses() {
 	});
 }
 
-export class AccessibilityHelper {
-	public static initA11YView(view: View): void {
-		if (view instanceof ProxyViewContainer) {
-			return;
-		}
-
-		const uiView = view.nativeViewProtected as UIView;
-		if (!uiView) {
-			return;
-		}
-
-		uiViewToTnsView.set(uiView, new WeakRef(view));
+export function initA11YView(view: View): void {
+	if (view instanceof ProxyViewContainer) {
+		return;
 	}
 
-	public static updateAccessibilityProperties(view: View): void {
-		if (view instanceof ProxyViewContainer) {
-			return;
-		}
-
-		const uiView = view.nativeViewProtected as UIView;
-		if (!uiView) {
-			return;
-		}
-
-		ensureNativeClasses();
-
-		const accessibilityRole = view.accessibilityRole;
-		const accessibilityState = view.accessibilityState;
-
-		if (!view.accessible || view.accessibilityHidden) {
-			uiView.accessibilityTraits = UIAccessibilityTraitNone;
-
-			return;
-		}
-
-		let a11yTraits = UIAccessibilityTraitNone;
-		if (RoleTypeMap.has(accessibilityRole)) {
-			a11yTraits |= RoleTypeMap.get(accessibilityRole);
-		}
-
-		switch (accessibilityRole) {
-			case AccessibilityRole.Checkbox:
-			case AccessibilityRole.RadioButton:
-			case AccessibilityRole.Switch: {
-				if (accessibilityState === AccessibilityState.Checked) {
-					a11yTraits |= AccessibilityTraitsMap.get(AccessibilityTrait.Selected);
-				}
-				break;
-			}
-			default: {
-				if (accessibilityState === AccessibilityState.Selected) {
-					a11yTraits |= AccessibilityTraitsMap.get(AccessibilityTrait.Selected);
-				}
-				if (accessibilityState === AccessibilityState.Disabled) {
-					a11yTraits |= AccessibilityTraitsMap.get(AccessibilityTrait.NotEnabled);
-				}
-				break;
-			}
-		}
-
-		const UpdatesFrequentlyTrait = AccessibilityTraitsMap.get(AccessibilityTrait.UpdatesFrequently);
-
-		switch (view.accessibilityLiveRegion) {
-			case AccessibilityLiveRegion.Polite:
-			case AccessibilityLiveRegion.Assertive: {
-				a11yTraits |= UpdatesFrequentlyTrait;
-				break;
-			}
-			default: {
-				a11yTraits &= ~UpdatesFrequentlyTrait;
-				break;
-			}
-		}
-
-		if (view.accessibilityMediaSession) {
-			a11yTraits |= AccessibilityTraitsMap.get(AccessibilityTrait.StartsMediaSession);
-		}
-
-		if (view.accessibilityTraits) {
-			a11yTraits |= inputArrayToBitMask(view.accessibilityTraits, AccessibilityTraitsMap);
-		}
-
-		uiView.accessibilityTraits = a11yTraits;
+	const uiView = view.nativeViewProtected as UIView;
+	if (!uiView) {
+		return;
 	}
 
-	public static sendAccessibilityEvent(view: View, eventName: string, text?: string): void {
-		// ignore
-	}
-
-	public static updateContentDescription(view: View, forceUpdate?: boolean): string {
-		return null;
-	}
-
-	public static isAccessibilityServiceEnabled(): boolean {
-		return isAccessibilityServiceEnabled();
-	}
+	uiViewToTnsView.set(uiView, new WeakRef(view));
 }
 
-let accessbilityServiceEnabled: boolean;
-let nativeObserver: any;
+export function updateAccessibilityProperties(view: View): void {
+	if (view instanceof ProxyViewContainer) {
+		return;
+	}
 
-function isAccessibilityServiceEnabled(): boolean {
+	const uiView = view.nativeViewProtected as UIView;
+	if (!uiView) {
+		return;
+	}
+
+	ensureNativeClasses();
+
+	const accessibilityRole = view.accessibilityRole;
+	const accessibilityState = view.accessibilityState;
+
+	if (!view.accessible || view.accessibilityHidden) {
+		uiView.accessibilityTraits = UIAccessibilityTraitNone;
+
+		return;
+	}
+
+	let a11yTraits = UIAccessibilityTraitNone;
+	if (RoleTypeMap.has(accessibilityRole)) {
+		a11yTraits |= RoleTypeMap.get(accessibilityRole);
+	}
+
+	switch (accessibilityRole) {
+		case AccessibilityRole.Checkbox:
+		case AccessibilityRole.RadioButton:
+		case AccessibilityRole.Switch: {
+			if (accessibilityState === AccessibilityState.Checked) {
+				a11yTraits |= AccessibilityTraitsMap.get(AccessibilityTrait.Selected);
+			}
+			break;
+		}
+		default: {
+			if (accessibilityState === AccessibilityState.Selected) {
+				a11yTraits |= AccessibilityTraitsMap.get(AccessibilityTrait.Selected);
+			}
+			if (accessibilityState === AccessibilityState.Disabled) {
+				a11yTraits |= AccessibilityTraitsMap.get(AccessibilityTrait.NotEnabled);
+			}
+			break;
+		}
+	}
+
+	const UpdatesFrequentlyTrait = AccessibilityTraitsMap.get(AccessibilityTrait.UpdatesFrequently);
+
+	switch (view.accessibilityLiveRegion) {
+		case AccessibilityLiveRegion.Polite:
+		case AccessibilityLiveRegion.Assertive: {
+			a11yTraits |= UpdatesFrequentlyTrait;
+			break;
+		}
+		default: {
+			a11yTraits &= ~UpdatesFrequentlyTrait;
+			break;
+		}
+	}
+
+	if (view.accessibilityMediaSession) {
+		a11yTraits |= AccessibilityTraitsMap.get(AccessibilityTrait.StartsMediaSession);
+	}
+
+	if (view.accessibilityTraits) {
+		a11yTraits |= inputArrayToBitMask(view.accessibilityTraits, AccessibilityTraitsMap);
+	}
+
+	uiView.accessibilityTraits = a11yTraits;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-empty-function
+export const sendAccessibilityEvent = (): void => {};
+export const updateContentDescription = (): string | null => null;
+
+let accessbilityServiceEnabled: boolean;
+let nativeObserver: NSNotification;
+export function isAccessibilityServiceEnabled(): boolean {
 	if (typeof accessbilityServiceEnabled === 'boolean') {
 		return accessbilityServiceEnabled;
 	}
