--- packages/core/acessibility/index.ts
@@ -1,3 +1,0 @@
-export * from './acessibility-properties';
-export * from './fontscale-observable';
-export * from './types';
