+++ packages/core/ui/core/view/index.d.ts
@@ -6,7 +6,7 @@ import { Animation, AnimationDefinition, AnimationPromise } from '../../animatio
 import { HorizontalAlignment, VerticalAlignment, Visibility, Length, PercentLength } from '../../styling/style-properties';
 import { GestureTypes, GestureEventData, GesturesObserver } from '../../gestures';
 import { LinearGradient } from '../../styling/gradient';
-import { AccessibilityLiveRegion, AccessibilityRole, AccessibilityState, AccessibilityTrait } from '../../../acessibility/types';
+import { AccessibilityLiveRegion, AccessibilityRole, AccessibilityState, AccessibilityTrait } from '../../../acessibility/accessibility-types';
 
 // helpers (these are okay re-exported here)
 export * from './view-helper';
